requirements.txt
streamlit>=1.36
openai>=1.40.0
pinecone>=5.0.0
pydantic>=2.5.0
pypdf>=4.0.0
python-docx>=1.0.0
pymongo>=4.7.3
dnspython>=2.6.1
